GenericPasswordConvertible for string
